fix compliant stats service method change type Int to long

+++ complaint-service/src/main/java/com/complaintService/complaintService/dto/complaintStatsDto.java
@@ -8,7 +8,7 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 @AllArgsConstructor
 public class complaintStatsDto {
-    private Integer totalComplaints;
+    private Long totalComplaints;
     private Integer resolvedComplaints;
     private Integer pendingComplaints;
     private Integer inProgressComplaints;
